+++ crates/scheduler/src/call/store/fragment.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2023 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,30 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use wasefire_applet_api::store::{self as api, Api};
+use core::ops::Range;
+
+use wasefire_applet_api::store::fragment::{self as api, Api};
 use wasefire_board_api::Api as Board;
-use wasefire_store::StoreError;
+use wasefire_store::fragment;
 
+use super::convert;
 use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
-
-mod fragment;
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::Insert(call) => insert(call),
         Api::Remove(call) => remove(call),
         Api::Find(call) => find(call),
-        Api::Fragment(call) => fragment::process(call),
     }
 }
 
 fn insert<B: Board>(mut call: SchedulerCall<B, api::insert::Sig>) {
-    let api::insert::Params { key, ptr, len } = call.read();
+    let api::insert::Params { keys, ptr, len } = call.read();
     let scheduler = call.scheduler();
     let memory = scheduler.applet.memory();
     let results = try {
+        let keys = decode_keys(keys)?;
         let value = memory.get(*ptr, *len)?;
-        let res = match scheduler.store.insert(*key as usize, value) {
+        let res = match fragment::write(&mut scheduler.store, &keys, value) {
             Ok(()) => 0.into(),
             Err(e) => convert(e).into(),
         };
@@ -45,24 +46,24 @@ fn insert<B: Board>(mut call: SchedulerCall<B, api::insert::Sig>) {
 }
 
 fn remove<B: Board>(mut call: SchedulerCall<B, api::remove::Sig>) {
-    let api::remove::Params { key } = call.read();
-    let res = match call.scheduler().store.remove(*key as usize) {
-        Ok(()) => 0.into(),
-        Err(e) => convert(e).into(),
+    let api::remove::Params { keys } = call.read();
+    let results = try {
+        let res = match fragment::delete(&mut call.scheduler().store, &decode_keys(keys)?) {
+            Ok(()) => 0.into(),
+            Err(e) => convert(e).into(),
+        };
+        api::remove::Results { res }
     };
-    call.reply(Ok(api::remove::Results { res }));
+    call.reply(results);
 }
 
 fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
-    #[cfg(feature = "multivalue")]
-    let api::find::Params { key } = call.read();
-    #[cfg(not(feature = "multivalue"))]
-    let api::find::Params { key, ptr: ptr_ptr, len: len_ptr } = call.read();
+    let api::find::Params { keys, ptr: ptr_ptr, len: len_ptr } = call.read();
     let scheduler = call.scheduler();
     let mut memory = scheduler.applet.memory();
     let results = try {
         let mut results = api::find::Results::default();
-        match scheduler.store.find(*key as usize) {
+        match fragment::read(&scheduler.store, &decode_keys(keys)?) {
             Ok(None) => (),
             Ok(Some(value)) => {
                 let len = value.len() as u32;
@@ -72,21 +73,10 @@ fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
                     Err(Trap)?;
                 }
                 memory.get_mut(ptr, len)?.copy_from_slice(&value);
-                #[cfg(feature = "multivalue")]
-                {
-                    results.ptr = ptr.into();
-                    results.len = len.into();
-                }
-                #[cfg(not(feature = "multivalue"))]
-                {
-                    memory.get_mut(*ptr_ptr, 4)?.copy_from_slice(&ptr.to_le_bytes());
-                    memory.get_mut(*len_ptr, 4)?.copy_from_slice(&len.to_le_bytes());
-                    results.res = 1.into();
-                }
+                memory.get_mut(*ptr_ptr, 4)?.copy_from_slice(&ptr.to_le_bytes());
+                memory.get_mut(*len_ptr, 4)?.copy_from_slice(&len.to_le_bytes());
+                results.res = 1.into();
             }
-            #[cfg(feature = "multivalue")]
-            Err(e) => results.len = convert(e).into(),
-            #[cfg(not(feature = "multivalue"))]
             Err(e) => results.res = convert(e).into(),
         }
         results
@@ -94,12 +84,10 @@ fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
     call.reply(results);
 }
 
-fn convert(err: StoreError) -> api::Error {
-    match err {
-        StoreError::InvalidArgument => api::Error::InvalidArgument,
-        StoreError::NoCapacity => api::Error::NoCapacity,
-        StoreError::NoLifetime => api::Error::NoLifetime,
-        StoreError::StorageError => api::Error::StorageError,
-        StoreError::InvalidStorage => api::Error::InvalidStorage,
+fn decode_keys(keys: u32) -> Result<Range<usize>, Trap> {
+    if keys & 0xf000f000 == 0 {
+        Ok((keys & 0xffff) as usize .. ((keys >> 16) & 0xffff) as usize)
+    } else {
+        Err(Trap)
     }
 }