+++ crates/scheduler/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Minor
 
 - Support `syscall()`
+- Support `store::fragment`
 
 ### Patch
 
@@ -61,4 +62,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 16 -->
+<!-- Increment to skip CHANGELOG.md test: 17 -->